+++ src/combinator.rs
@@ -118,11 +118,13 @@ impl<T> Combinator<T> {
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>> Action<State, Heap> for Combinator<T> {
+unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap>> Action<Text, State, Heap>
+  for Combinator<T>
+{
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, input: Input<&str, &mut State, &mut Heap>) -> Option<Output<T::Value>> {
+  fn exec(&self, input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<T::Value>> {
     self.action.exec(input)
   }
 }