+++ src/combinator/decorator/state.rs
@@ -6,42 +6,53 @@ create_closure_decorator!(Then, "See [`Combinator::then`].");
 create_closure_decorator!(Catch, "See [`Combinator::catch`].");
 create_closure_decorator!(Finally, "See [`Combinator::finally`].");
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
-  Action<State, Heap> for Prepare<T, D>
+unsafe impl<
+    Text: ?Sized,
+    State,
+    Heap,
+    T: Action<Text, State, Heap>,
+    D: Fn(Input<&Text, &mut State, &mut Heap>),
+  > Action<Text, State, Heap> for Prepare<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     (self.inner)(input.reborrow());
     self.action.exec(input)
   }
 }
 
 unsafe impl<
+    Text: ?Sized,
     State,
     Heap,
-    T: Action<State, Heap>,
-    D: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>),
-  > Action<State, Heap> for Then<T, D>
+    T: Action<Text, State, Heap>,
+    D: Fn(AcceptedContext<Input<&Text, &mut State, &mut Heap>, &Output<T::Value>>),
+  > Action<Text, State, Heap> for Then<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).inspect(|output| {
       (self.inner)(AcceptedContext { input, output });
     })
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
-  Action<State, Heap> for Catch<T, D>
+unsafe impl<
+    Text: ?Sized,
+    State,
+    Heap,
+    T: Action<Text, State, Heap>,
+    D: Fn(Input<&Text, &mut State, &mut Heap>),
+  > Action<Text, State, Heap> for Catch<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     let output = self.action.exec(input.reborrow());
     if output.is_none() {
       (self.inner)(input);
@@ -50,13 +61,18 @@ unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &
   }
 }
 
-unsafe impl<State, Heap, T: Action<State, Heap>, D: Fn(Input<&str, &mut State, &mut Heap>)>
-  Action<State, Heap> for Finally<T, D>
+unsafe impl<
+    Text: ?Sized,
+    State,
+    Heap,
+    T: Action<Text, State, Heap>,
+    D: Fn(Input<&Text, &mut State, &mut Heap>),
+  > Action<Text, State, Heap> for Finally<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     let output = self.action.exec(input.reborrow());
     (self.inner)(input);
     output
@@ -75,12 +91,12 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn prepare<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
+  pub fn prepare<Text: ?Sized, State, Heap, F: Fn(Input<&Text, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Prepare<T, F>>
   where
-    T: Action<State, Heap>,
+    T: Action<Text, State, Heap>,
   {
     Combinator::new(Prepare::new(self.action, modifier))
   }
@@ -97,15 +113,16 @@ impl<T> Combinator<T> {
   /// ```
   #[inline]
   pub fn then<
+    Text: ?Sized,
     State,
     Heap,
-    F: Fn(AcceptedContext<Input<&str, &mut State, &mut Heap>, &Output<T::Value>>),
+    F: Fn(AcceptedContext<Input<&Text, &mut State, &mut Heap>, &Output<T::Value>>),
   >(
     self,
     modifier: F,
   ) -> Combinator<Then<T, F>>
   where
-    T: Action<State, Heap>,
+    T: Action<Text, State, Heap>,
   {
     Combinator::new(Then::new(self.action, modifier))
   }
@@ -121,12 +138,12 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn catch<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
+  pub fn catch<Text: ?Sized, State, Heap, F: Fn(Input<&Text, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Catch<T, F>>
   where
-    T: Action<State, Heap>,
+    T: Action<Text, State, Heap>,
   {
     Combinator::new(Catch::new(self.action, modifier))
   }
@@ -143,12 +160,12 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn finally<State, Heap, F: Fn(Input<&str, &mut State, &mut Heap>)>(
+  pub fn finally<Text: ?Sized, State, Heap, F: Fn(Input<&Text, &mut State, &mut Heap>)>(
     self,
     modifier: F,
   ) -> Combinator<Finally<T, F>>
   where
-    T: Action<State, Heap>,
+    T: Action<Text, State, Heap>,
   {
     Combinator::new(Finally::new(self.action, modifier))
   }
@@ -165,14 +182,14 @@ mod tests {
     to: i32,
   }
 
-  fn accepter() -> Combinator<impl Action<State, Value = ()>> {
+  fn accepter() -> Combinator<impl Action<str, State, Value = ()>> {
     wrap(|input: Input<&str, &mut State, &mut ()>| {
       input.state.to = input.state.from;
       input.digest(1)
     })
   }
 
-  fn rejecter() -> Combinator<impl Action<State, Value = ()>> {
+  fn rejecter() -> Combinator<impl Action<str, State, Value = ()>> {
     wrap(|input: Input<&str, &mut State, &mut ()>| {
       input.state.to = input.state.from;
       None