+++ src/combinator/ops/mul/fold.rs
@@ -2,6 +2,7 @@ use super::{Mul, Repeat};
 use crate::{
   action::{Action, Input, Output},
   combinator::{ops::mul::impl_mul, Combinator},
+  digest::Digest,
 };
 use std::ops;
 
@@ -10,18 +11,18 @@ use std::ops;
 /// See [`ops::mul`](crate::combinator::ops::mul) for more information.
 ///
 /// Built-in implementations are provided for `()`.
-pub trait Fold<State = (), Heap = ()> {
+pub trait Fold<Text: ?Sized = str, State = (), Heap = ()> {
   /// The accumulator type.
   type Output: Default;
 
   /// Fold self with the accumulator.
-  fn fold(self, acc: Self::Output, input: Input<&str, &mut State, &mut Heap>) -> Self::Output;
+  fn fold(self, acc: Self::Output, input: Input<&Text, &mut State, &mut Heap>) -> Self::Output;
 }
 
-impl<State, Heap> Fold<State, Heap> for () {
+impl<Text: ?Sized, State, Heap> Fold<Text, State, Heap> for () {
   type Output = ();
   #[inline]
-  fn fold(self, _: Self::Output, _: Input<&str, &mut State, &mut Heap>) -> Self::Output {}
+  fn fold(self, _: Self::Output, _: Input<&Text, &mut State, &mut Heap>) -> Self::Output {}
 }
 
 impl<Lhs, Rhs: Repeat> ops::Mul<Rhs> for Combinator<Lhs> {
@@ -34,13 +35,20 @@ impl<Lhs, Rhs: Repeat> ops::Mul<Rhs> for Combinator<Lhs> {
   }
 }
 
-unsafe impl<State, Heap, Lhs: Action<State, Heap, Value: Fold<State, Heap>>, Rhs: Repeat>
-  Action<State, Heap> for Mul<Combinator<Lhs>, Rhs>
+unsafe impl<
+    Text: ?Sized,
+    State,
+    Heap,
+    Lhs: Action<Text, State, Heap, Value: Fold<Text, State, Heap>>,
+    Rhs: Repeat,
+  > Action<Text, State, Heap> for Mul<Combinator<Lhs>, Rhs>
+where
+  for<'a> &'a Text: Digest,
 {
-  type Value = <Lhs::Value as Fold<State, Heap>>::Output;
+  type Value = <Lhs::Value as Fold<Text, State, Heap>>::Output;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     impl_mul!(input, self.rhs, Default::default, Fold::fold, self.lhs)
   }
 }
@@ -61,7 +69,7 @@ mod tests {
 
   #[derive(Debug)]
   struct MyValue(usize);
-  impl<State, Heap> Fold<State, Heap> for MyValue {
+  impl<State, Heap> Fold<str, State, Heap> for MyValue {
     type Output = usize;
     fn fold(self, current: Self::Output, _: Input<&str, &mut State, &mut Heap>) -> Self::Output {
       self.0 + current