+++ src/parser/builder.rs
@@ -73,7 +73,10 @@ impl<T, State, Heap> Builder<T, State, Heap> {
 
   /// Set [`Parser::entry`].
   #[inline]
-  pub fn entry<Entry: Action<State, Heap>>(self, entry: Entry) -> Builder<Entry, State, Heap> {
+  pub fn entry<Text: ?Sized, Entry: Action<Text, State, Heap>>(
+    self,
+    entry: Entry,
+  ) -> Builder<Entry, State, Heap> {
     Builder {
       entry,
       state: self.state,
@@ -82,10 +85,13 @@ impl<T, State, Heap> Builder<T, State, Heap> {
   }
 }
 
-impl<T: Action<State, Heap>, State, Heap> Builder<T, State, Heap> {
+impl<T, State, Heap> Builder<T, State, Heap> {
   /// Build a [`Parser`] with the given text.
   #[inline]
-  pub fn build(self, text: &str) -> Parser<T, State, Heap> {
+  pub fn build<Text: ?Sized>(self, text: &Text) -> Parser<T, Text, State, Heap>
+  where
+    T: Action<Text, State, Heap>,
+  {
     Parser {
       state: self.state,
       heap: self.heap,